Add example of deleting an array element.
Reading this doc made me think it wasn't including the situation where `delete` was used to remove an array element - which does not update the array's length which is often unexpected. However, trying it, it does flag this as an error, so I think it would be useful to have it in the docs.

```
  3:1  error  Modifying an existing object/array is not allowed  functional/immutable-data
```

Maybe it could use to more explanation, but I'm not sure where that would do in this document.