feat(prefer-immutable-types): change the options in recommeneded and lite configs
Recommended: Parameters must be deeply readonly when typed with a type defined in a local file.
Types defined in packages in the ts lib must be readonly shallow.

Lite: Parameters must be shallowly readonly when typed with a type defined in a local file.
There is no enforcement on types defined elsewhere.